fastapi_microservice/pyproject.toml
[project]
name = "fastapi_microservice"
version = "0.1.0"
description = "Microservice to load Meter data into cassandra DB"
authors = [
    {name = "madan", email = "mchalla.dev@live.com"},
]
dependencies = [
    "fastapi>=0.112.1",
    "uvicorn>=0.30.6",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.9.2",
    "cassandra-driver>=3.29.2",
    "prettytable>=3.11.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["src"]
